refactor(cli): move the oauth module out of the command namespace
`src/commands/oauth/` holds no commander command: `oauth setup` is gone and
nothing registers an `oauth` command. It is auth support code that the login,
logout, session and program modules depend on, so it belongs next to the other
shared modules in `src/lib/`.

Pure move plus import paths; no behavior change.

Fixes DX-490
