+++ packages/cli/src/lib/oauth/space-guard.test.ts
