+++ packages/cli/src/lib/oauth/space-guard.ts
